--- src/pages/[locale]/index.tsx
@@ -1,1 +1,0 @@
-export { default, getServerSideProps } from 'pages/[locale]/[slug]';
