+++ src/pages/_app.tsx
@@ -12,7 +12,7 @@ function HelpdeskApp({ Component, pageProps }) {
 
   return (
     <LocaleContext.Provider value={locale}>
-      <div className="flex flex-col bg-white">
+      <div>
         <Head>
           <link rel="shortcut icon" href="/favicon/favicon.png" type="image/png" />
           <link rel="apple-touch-icon" href="/favicon/apple-icon-57x57.png" type="image/png" />
@@ -20,12 +20,12 @@ function HelpdeskApp({ Component, pageProps }) {
           <link rel="apple-touch-icon" href="/favicon/apple-icon-114x114.png" type="image/png" />
           <meta
             name="description"
-            content={`Demo Help Center built using Next.js and Contentful Compose.`}
+            content={`Demo using Next.js and Contentful Compose.`}
             key="description"
           />
         </Head>
         <PreviewBanner />
-        <div className="w-full flex flex-col relative">
+        <div>
           <TopNavigation />
           <Component {...otherPageProps} />
         </div>
@@ -33,20 +33,5 @@ function HelpdeskApp({ Component, pageProps }) {
     </LocaleContext.Provider>
   );
 }
-
-HelpdeskApp.getInitialProps = async (appContext: AppContext) => {
-  const { pageProps, ...other } = await App.getInitialProps(appContext);
-  const locale = getLocale(appContext.ctx.query.locale);
-
-  switch (locale) {
-    case UnknownLocale:
-      appContext.ctx.res.writeHead(302, { Location: '/' }).end();
-      break;
-    case undefined:
-      return { pageProps, ...other };
-    default:
-      return { pageProps: { ...pageProps, locale }, ...other };
-  }
-};
 
 export default HelpdeskApp;
